3단계 - 장바구니(상품 담기) (#42)
* step3 구현사항 정리

* step2 리뷰 반영

* feat: 장바구니 UI 구현

* feat: Cart Object 추가

* feat: Cart 클릭과 관련된 비즈니스 로직 구현

* feat: 총 주문 금액 구현

* test: 상품 갯수 테스트 코드 추가

* test: 장바구니 총 주문 금액 테스트 코드 추가

* test: 장바구니 아이템 clear 버튼 검증

* docs: readme 정리

* test: 코드 순서 및 정리

+++ app/src/main/java/nextstep/shoppingcart/ui/screen/products/model/ProductModel.kt
@@ -8,7 +8,7 @@ data class ProductModel(
     val id: String = UUID.randomUUID().toString(),
     val name: String,
     val imageUrl: String,
-    val price: Long,
+    val price: Int,
 )
 
 val dummyProductModels: List<ProductModel> = listOf(
